Create SAS-SQL Example 1
This code was written in SAS environment however it should work with in any SQL environment.  With this code you will be able to to track problem types, Heart Rate, SBP, and DBP. Based on data input as a number line. The program will calculate the means of HR,SBP,DBP. As well as the frequency of the other problems for patents.